+++ packages/cadl-ranch-specs/http/type/scalar/main.tsp
@@ -61,3 +61,62 @@ interface Unknown {
   @doc("put unknown value")
   put(@body @doc("_") body: unknown): void;
 }
+
+@doc("Template to have scalar types operations")
+interface ScalarTypesOperations<T, TDoc extends string> {
+  @scenario
+  @scenarioDoc(
+    """
+  Expected response body:
+  ```json
+  {doc}
+  ```
+  """,
+    {
+      doc: TDoc,
+    }
+  )
+  @get
+  @route("/response_body")
+  responseBody(): T;
+
+  @scenario
+  @scenarioDoc(
+    """
+  Expected input body:
+  ```json
+  {doc}
+  ```
+  """,
+    {
+      doc: TDoc,
+    }
+  )
+  @put
+  @route("/resquest_body")
+  requestBody(@body body: T): void;
+
+  @scenario
+  @scenarioDoc(
+    """
+  Expected request parameter:
+  value={doc}
+  """,
+    {
+      doc: TDoc,
+    }
+  )
+  @get
+  @route("/request_parameter")
+  requestParameter(@query value: T): void;
+}
+
+@doc("Decimal type")
+@route("/decimal")
+@operationGroup
+interface DecimalType extends ScalarTypesOperations<decimal, "0.33333"> {}
+
+@doc("Decimal128 type")
+@route("/decimal128")
+@operationGroup
+interface Decimal128Type extends ScalarTypesOperations<decimal128, "0.33333"> {}
